==> docs/Notebooks/Nb2_Crime_Rates_UK.ipynb <==
# Investigating Crime Rates UK

We found an API that allows you to put the longitude and latitude of any area and it returned the crime data in a mile radius and date= #https://data.police.uk/docs/method/crime-street/

**We used this to investigate:**

1. effect of covid lockdowns on crime rates in cities across the UK =  Hypothesis was that at the start of 2022, when covid restrictions were most intense, crime rates would be the lowest

2. Most Common Crimes in London by Category

import requests
import pandas as pd
import streamlit as st
import altair as alt

# 1.Effect of COVID lockdowns on crime rates 


### Newcastle Central Station 

# create list with months of the year 
months=['01', '02', '03', '04', '05', '06', '07', '08', '09', '10', '11', '12']


#create emprty list
crimerates2021newcastle=[]

#loop through the months in the year in the URL
for month in months:
    url = 'https://data.police.uk/api/crimes-street/all-crime?lat=54.9691552&lng=-1.6188458&date=2021-'+month+''
    response = requests.get(url)
    if response.status_code == 200:
        data = response.json() 
        crimerates2021newcastle.append(len(data))
    else:
        print(f"Error: {response.status_code}")
        print(response.text)


print(crimerates2021newcastle)


 ### London Kings Cross

# repeated above process for the following 3 cities
import requests

months=['01', '02', '03', '04', '05', '06', '07', '08', '09', '10', '11', '12']

crimerates2021london=[]

for month in months:
    url = 'https://data.police.uk/api/crimes-street/all-crime?lat=51.5307106&lng=-0.1274346&date=2021-'+month+''
    response = requests.get(url)
    if response.status_code == 200:
        data = response.json()  # Assuming the response is in JSON format
        crimerates2021london.append(len(data))
    else:
        print(f"Error: {response.status_code}")
        print(response.text)

print(crimerates2021london)

### Manchester Piccadilly Station

months=['01', '02', '03', '04', '05', '06', '07', '08', '09', '10', '11', '12']

crimerates2021manchester=[]

for month in months:
    url = 'https://data.police.uk/api/crimes-street/all-crime?lat=53.4810422&lng=-2.2440995&date=2021-'+month+''
    response = requests.get(url)
    if response.status_code == 200:
        data = response.json()  # Assuming the response is in JSON format
        crimerates2021manchester.append(len(data))
    else:
        print(f"Error: {response.status_code}")
        print(response.text)


print(crimerates2021manchester)



### Portsmouth Fratton Station

months=['01', '02', '03', '04', '05', '06', '07', '08', '09', '10', '11', '12']

crimerates2021portsmouth=[]

for month in months:
    url = 'https://data.police.uk/api/crimes-street/all-crime?lat=50.7964591&lng=-1.0764905&date=2021-'+month+''
    response = requests.get(url)
    if response.status_code == 200:
        data = response.json()  # Assuming the response is in JSON format
        crimerates2021portsmouth.append(len(data))
    else:
        print(f"Error: {response.status_code}")
        print(response.text)


print(crimerates2021portsmouth)


# create a data frame with crime rates of each UK city 

UKcrimerates2021 = pd.DataFrame({
    'Month': months,
    'London': crimerates2021london,
    'Manchester': crimerates2021manchester,
    'Newcastle': crimerates2021newcastle,
    'Portsmouth': crimerates2021portsmouth
})


crimerates2021.to_csv('..Data/crimerates2021.csv', index=False)

### Creating Crime Rate graph

# creating a graph using altaire, chatGBT helped us with this as we were unsure how to add multiple lines to graph

UKcrimerategraph = alt.Chart(crimerates2021).mark_line().encode(
    x=alt.X('Month:O', title='Month'),  
    color=alt.Color('city:N', title='City'), 
    y=alt.Y('crime_rate:Q', title='Crime Rates'),
    tooltip=['Month:O', 'city:N', 'crime_rate:Q'] 
).properties(
    width=500, 
    height=300,
    title='Crime Rates in Cities (2021)'
).transform_fold(  
    fold=['London', 'Manchester', 'Newcastle', 'Portsmouth' ],
    as_=['city', 'crime_rate']

).interactive()


UKcrimerategraph 


Manchester results made us question the collection of data- problems with effiecency of data collection in different cities. Although manchester had slightly more intense lockdowns, it is unlikely this caused such an intense difference to newcastle, as they are similar sizes.


# 2. Exploring Most Common Crimes in London by Category

#sending request to the URL
url = 'https://data.police.uk/api/crimes-street/all-crime?lat=54.991677&lng=-1.603831&date=2022-01'
response = requests.get(url)
data = response.json()

#creating list of months to loop through
months=['01', '02', '03', '04', '05', '06', '07', '08', '09', '10', '11', '12']

#creating empty list
crimerateslondon=[]

for month in months:
    url = 'https://data.police.uk/api/crimes-street/all-crime?lat=51.5307106&lng=-0.1274346&date=2021-'+month+''
    response = requests.get(url)
    if response.status_code == 200:
        data = response.json()  

# creating empty dictionary and month list
months = ['01', '02', '03', '04', '05', '06', '07', '08', '09', '10', '11', '12']
crimelondon = []

#looping through months and saving all data for each crime to 'crimelondon'
for month in months:
    url = f'https://data.police.uk/api/crimes-street/all-crime?lat=51.5307106&lng=-0.1274346&date=2021-{month}'
    response = requests.get(url)
    
    if response.status_code == 200:
        data = response.json()
        crimelondon.extend(data)

# turning crimelondon into a dataframe
londoncrimeinformation = pd.DataFrame(crimelondon)

londoncrimeinformation.head()

### Cleaning Data

Above turned 26,000 rows of data with alot of irrelevent information eg crime id. Below we cleaned the data so that we just had relevent data, then used value_counts() to count crimes per category

#cleaning the data so that we just had relevent data, then used value_counts() to count crimes per category

crimecategorylondon2022 = londoncrimeinformation['category'].value_counts().reset_index()
crimecategorylondon2022.columns = ['category', 'count']

# saving to csv in data folder
crimecategorylondon2022.to_csv('..Data/crimecategorylondon2022.csv', index=False)

crimecategorylondon2022


### creating bar chart

# creating graph- used ChatGBT to help us with this and had only made line graphs in altaire previously

Category_bar_chart = alt.Chart(crimecategorylondon2022).mark_bar().encode(
    x='count:Q',
    y=alt.Y('category:N', sort='-x'),
    tooltip=['category:N', 'count:Q'] 
).properties(
    title='Crime by Category in London 2021' 
)


Category_bar_chart
