==> docs/Notebooks/Nb3_Global_Average_GINI.ipynb <==
import pandas as pd

data contains gini for each country and each year, needs to be condensed to show average gini for every country over each year

gini_df = pd.read_csv("../Data/Do_Not_Edit/UNdata_Export.csv")
gini_df.head()

Here values are grouped by each year and an average global gini is created from the mean

avg_global_gini_df = gini_df.groupby(['Year'])['Value'].mean().reset_index()
avg_global_gini_df['Year'] = pd.to_numeric(avg_global_gini_df['Year'], errors='coerce') 
avg_global_gini_df = avg_global_gini_df.dropna(subset=['Year']) 
avg_global_gini_df.columns = ["Year","GlobalGini"]
avg_global_gini_df.head()

Saving as CSV

avg_global_gini_df.to_csv("../Data/AverageGlobalGini.csv")
