==> docs/Notebooks/Nb5_GlobalPop.ipynb <==
# Global Population Predictor Data
This data was scraped in order to predict global populations, separated by country, which shows the predicted data for the year 2100. This portion is added to the global data found to add.

import requests
import re
import pandas as pd

#Importing commands

url = "https://www.ined.fr/en/everything_about_population/data/world-projections/projections-by-countries/"
response = requests.get(url)
#Providing the URL for data extraction
if response.status_code == 200:
    html_content = response.text
    #Dividing the columns in the table
    table_pattern = re.compile(r'<table[^>]*>.*?</table>', re.DOTALL)
    table_match = table_pattern.search(html_content)

    if table_match:
        table_html = table_match.group(0)

        header_pattern = re.compile(r'<th.*?>(.*?)</th>', re.DOTALL)
        variables = header_pattern.findall(table_html)

        cell_pattern = re.compile(r'<td.*?>(.*?)</td>', re.DOTALL)
        cells = cell_pattern.findall(table_html)
        #Arranging the length of variables in the table for cohesion
        if variables and cells:
            data = [cells[i:i + len(variables)] for i in range(0, len(cells), len(variables))]
            df = pd.DataFrame(data, columns=variables)
            
        else:
            print("Variable names or data not found in the table.")
    else:
        print("Table not found.")
else:
    print("Failed to retrieve the webpage. Status code:", response.status_code)
df = df.drop(df.columns[[0,1]], axis=1)
# converting columns to numeric
for col in ['2050', '2075', '2100']:
    # Data has non standard space breaks which is what the str.replace accounts for
    df[col] = df[col].str.replace('\u202f', '').astype(float)
df

#Help from ChatGPT to edit the code was utilised when the code written didn't provide output that was divided into columns


simplifying data to show population by region

region_pop_df = df.groupby('Region').sum()
region_pop_df=region_pop_df[:-1]
region_pop_df

Generic pie chart method

import altair as alt

def create_pie_chart(year):
    # creating data frame for single year
    
    data_year = region_pop_df.reset_index()[['Region', year]].rename(columns={year: 'Population'})
    chart = alt.Chart(data_year).mark_arc().encode(
        theta=alt.Theta(field="Population", type="quantitative"),
        color=alt.Color(field="Region", type="nominal"),
        tooltip=['Region', 'Population']
    ).properties(
        title=f"Population Distribution in {year}"
    )
    return chart

chart_2050 = create_pie_chart('2050')
chart_2075 = create_pie_chart('2075')
chart_2100 = create_pie_chart('2100')


chart_2050 | chart_2075 | chart_2100

Saving CSV

region_pop_df.to_csv("../Data/Globalpopulation.csv")
