==> docs/Notebooks/Nb6_UK_Voter_Turnout.ipynb <==
# UK Voter Turnout Data
This data summarises the yearly voter turnout for each election. The columns were cleaned accordingly, to only represent the first four numbers for the year column, and the voter turnout in percentage for the other column.

import pandas as pd
import altair as alt
from altair import datum

#Importing all the commands to utilise throughout


url = "https://www.theguardian.com/news/datablog/2010/may/06/general-election-2010-turnout-since-1945"
tables = pd.read_html(url)
table = tables[0]
print(table)
#Scraping the first table from the given URL


tables = pd.read_html(url)

df = tables[0][['Year', 'UK']].copy() 
#Creating a new dataframe with only selected columns
df.columns = ["Year", "Voter Turnout in UK"]
df['Year'] = df['Year'].str[:4]

df.loc[:, "Voter Turnout in UK"] = pd.to_numeric(df["Voter Turnout in UK"], errors="coerce")
#Creating an altair graph
chart = alt.Chart(df).mark_line().encode(
    x='Year:O',
    y='Voter Turnout in UK:Q',
    tooltip=['Year:O', 'Voter Turnout in UK:Q']
).properties(
    title='Voter Turnout in the UK - General Elections Since 1945'
)

#Turning data into a CSV file
df.to_csv('../Data/voter_turnout_data.csv', index=False)
df
