==> docs/Notebooks/Nb8_USA_yealy_gdpcapita.ipynb <==
# USA Yearly GDP Per Capita Data
This data summarises the GDP per capita in the USA for each year. GDP per capita was favoured insted of GDP as a better comparison between variables would be seen with GDP per capita.

import pandas as pd
import requests
#Importing commands


url = "https://www.macrotrends.net/countries/USA/united-states/gdp-per-capita"

headers = {
    'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36'
}
#Defining the user agent to avoid the blockage on the website
response = requests.get(url, headers=headers)

if response.status_code == 200:
    dfs = pd.read_html(response.text)
    
    if dfs:
        df = dfs[1]
        #Scraping the variables from the second table in the link
        print(df)
    else:
        print("No tables found")
else:
    print(f"Failed to retrieve the page. Status code: {response.status_code}")
#Saving the data to a new dataframe to only have two of the selected columns
df = df.rename(columns={'GDP Per Capita (US $)': 'USA GDP Per Capita (US $)'})
df



df.to_csv('usa_yearly_gdpcapita', index=False)
#Saving the data to a CSV file
