==> docs/Notebooks/Nb9_USA_yearly_inflation.ipynb <==
# USA Yearly Inflation Rates Data
This data summarises the USA inflation rates by year, where the total increase in percentage is seen, and per month division is visible.

import pandas as pd

#Importing commands


url = "https://www.usinflationcalculator.com/inflation/current-inflation-rates/"
#Scraping the first table in the link and saving to a dataframe
df_list = pd.read_html(url)
if df_list:
    df = df_list[0]
    
    print(df)
else:
    print("No tables found")
df 

df['Average'] = df.iloc[:, 1:-1].mean(axis=1, skipna=True)
df

df.to_csv("../Data/usa_yearly_inflation_rates.csv", index=False)
#Saving the data to a CSV file
