==> docs/Notebooks/Nb10_Interactive_graph.ipynb <==
import pandas as pd

avg_global_gini_df = pd.read_csv("../Data/AverageGlobalGini.csv")
uk_yearly_gdp_df = pd.read_csv("../Data/uk_yearly_gdpcapita.csv")
us_yearly_gdp_df = pd.read_csv("../Data/usa_yearly_gdpcapita.csv")
us_poverty_df = pd.read_csv("../Data/poverty.csv")
uk_voter_turnout_df = pd.read_csv("../Data/voter_turnout_data.csv")

Correcting column names

uk_yearly_gdp_df.columns = uk_yearly_gdp_df.iloc[0]
us_yearly_gdp_df.columns = us_yearly_gdp_df.iloc[0]
uk_yearly_gdp_df = uk_yearly_gdp_df.iloc[1:]
us_yearly_gdp_df = us_yearly_gdp_df.iloc[1:]

Correcting column datatypes to be the same in order to merge

avg_global_gini_df["Year"] = avg_global_gini_df["Year"].astype(int)
uk_yearly_gdp_df["Year"] = uk_yearly_gdp_df["Year"].astype(int)
us_yearly_gdp_df["Year"] = us_yearly_gdp_df["Year"].astype(int)

Removing unnecessary columns and renaming columns so they make sense in context of merged dataframe

del(avg_global_gini_df["Unnamed: 0"])
us_poverty_df.rename(columns={'Persons Below Poverty':'Persons Below Poverty (US)',	'Percent Below Poverty': 'Percent Below Poverty (US)'}, inplace=True)
uk_yearly_gdp_df.rename(columns={'Annual Growth Rate (%)': 'UK GDP Growth Rate'}, inplace=True)
us_yearly_gdp_df.rename(columns={'Annual Growth Rate (%)': 'US GDP Growth Rate'}, inplace=True)
avg_global_gini_df.rename(columns={'GlobalGini': 'Average Global Gini'}, inplace=True)

interactive_df = pd.merge(avg_global_gini_df, uk_yearly_gdp_df, on='Year', how='outer')
interactive_df = pd.merge(interactive_df, us_yearly_gdp_df, on='Year', how='outer')
interactive_df = pd.merge(interactive_df, us_poverty_df, on='Year', how='outer')
interactive_df = pd.merge(interactive_df, uk_voter_turnout_df, on='Year', how='outer')
interactive_df

Final clean of data removing dollar signs, percentage points and commas then converting to correct datatype

for column in interactive_df.columns:
    # skipping column that have non string contents
    if interactive_df[column].dtype == 'O':
        interactive_df[column] = interactive_df[column].str.replace('[$,%]', '', regex=True).astype(float)
    interactive_df[column] = pd.to_numeric(interactive_df[column], errors='coerce')
  
interactive_df

import altair as alt

Generic method for creating chart based of two columns

def plot_chart(df,col_x,col_y):
    chart = alt.Chart(df).mark_circle().encode(
        x=col_x,
        y=col_y,
        tooltip=['Year', col_x, col_y]
    )

    return chart

plot_chart(interactive_df,'Year','USA GDP Per Capita (US $)')

interactive_df.to_csv("../Data/InteractiveData.csv")
