==> docs/Notebooks/Nb11_Create_Database.ipynb <==
# Run this code after all other notebooks have been run

import sqlalchemy
from sqlalchemy import create_engine
import pandas as pd

loading sql package into jupyter notebook

%load_ext sql
%config SqlMagic.autocommit=True

Creating database connection

from sqlalchemy import create_engine
engine = create_engine('sqlite:///../Data/PoliticsandDataSci.db', echo=False, isolation_level="AUTOCOMMIT")
with engine.connect() as conn:
    pass

adding state data table

statedata_df = pd.read_csv('../Data/StateData.csv')
statedata_df.to_sql('State Data', con=engine, if_exists='replace', index=False)

adding biden opinion poll data

biden_opinion_df = pd.read_csv('../Data/Biden_Opinion_Polls_df.csv')
biden_opinion_df.to_sql('Biden polls', con=engine, if_exists='replace', index=False)
events_df = pd.read_csv("../Data/event_data.csv")
events_df.to_sql("Biden Events", con=engine, if_exists='replace', index=False)

Adding data for interactive graph

interactive_df = pd.read_csv("../Data/InteractiveData.csv")
interactive_df.to_sql('Interactive Data', con=engine, if_exists='replace', index=False)

Adding UK crime rate data

ukcrimerates_df = pd.read_csv("../Data/crimerates2021.csv")
ukcrimerates_df.to_sql('UK crime rates', con=engine, if_exists='replace', index=False)
londoncrimecategory_df = pd.read_csv("../Data/crimecategorylondon2022.csv")
londoncrimecategory_df.to_sql('London crime categories', con=engine, if_exists='replace', index=False)

Adding global population data

global_pop_df = pd.read_csv("../Data/Globalpopulation.csv")
global_pop_df.to_sql('Global population', con=engine, if_exists='replace', index=False)

Adding gini coefficient data

gini_df = pd.read_csv("../Data/AverageGlobalGini.csv")
gini_df.to_sql('Global gini', con=engine, if_exists='replace', index=False)
